webdav: interpret 'dcache-view.endpoints.webdav' as an override
Motivation:

The admin may configure a specific WebDAV endpoint that should be used
for all data transfers, and for the URL returned to the user (via a
context menu).  This is done with the `dcache-view.endpoints.webdav` (in
configuration, which is exposed in dCache as the
`frontend.static!dcache-view.endpoints.webdav` configuration property.

Currently, 'dcache-view.endpoints.webdav' only takes effect if there are
no auto-discovered doors; i.e., it is a fall-back.

This isn't terribly useful -- a more useful behaviour would be as an
override.  Like this the value is used, if supplied by the admin,
otherwise dCacheView will use a WebDAV door that it has discovered.

Modification:

Change the order of precedence: `dcache-view.endpoints.webdav` now has
the highest precedence.  If the admin specifies this, then it overrides
the auto-discovered value.

Result:

The admin has more control over which WebDAV door dCacheView uses.

Target: master
Request: 1.6
Request: 1.5
Closes: #252
Patch: https://rb.dcache.org/r/12976/
Acked-by: Tigran Mkrtchyan