oidc: add support for extra fields in OIDC (implicit flow)
Motivation:

Some OPs support extra fields in addition to those defined in OIDC.  As
a specific example, KeyCloak has a feature called "IdP hinting" where
the client (dCacheView) chooses how the user should authenticate to
KeyCloak.  This improves the login experience by removing a source of
possible mistakes and reducing the number of required clicks for the
user.

Therefore, it is helpful to allow the dCacheView configuration to
support additional key-value pairs that dCacheView will include in the
URL.

Modification:

Add support for an extra configuration option
'dcache-view.oidc-authz-endpoint-extra'.

Result:

dCacheView now supports an extra configuration option:
dcache-view.oidc-authz-endpoint-extra.  This is a space separated list
of material to be added during the OIDC login.  An empty or missing
value results in no extra material.  A single `-` character is a
placeholder for an empty string.