don't show login prompt for a macaroon link.
Motivation:

dCacheView always attempts to list the root directory when it starts.
If the user is not entitled to list the directory then dCacheView will
immediately prompt the user to log in.

If dCacheView is started with a macaroon then the macaroon is "consumed"
and becomes part of the "shared with me" selection.  This "shared with
me" view is selected automatically, so the view is NOT the filesystem
view.

However, despite this, dCacheView will still attempt to initiate the
file-system view.  If the door's root directory is not publicly listable
then this triggers a login request, which obscures the "shared with me"
view.

While it is possible to navigate back to the "shared with me" view and
view the contents made available via the macaroon, the navigation is not
obvious.  Moreover the prompt to log in is confusing: why prompt the
user to log in when they have been given a macaroon.

Modification:

The problem stems from the initialisation code.  The
'WebComponentsReady' event is triggered once all the components have
been loaded.

dCacheView currently has two functions triggered by this event: one
initialises the page routine while the other initialises the file-view
if the page is on the file-system view.  However, the order of these two
two functions is not defined, so it could be that the page routing
function is called after the file-view initialisation.

The current code has a work-around for this eventuality, by always
initialising the file-system view if the page's view is undefined;
however, this does not work if the page view is "shared with me".

This is fixed by moving all the 'WebComponentsReady' into a single
method.  The filesystem initialisation is now done within the method
called from within the routine change call-back function.

Result:

Following a macaroon link now shows the "shared with me" view, without
triggering a background filesytem root-directory listing, which might
the page showing the login prompt.

Target: master
Target: 1.6
Patch: https://rb.dcache.org/r/13055/
Acked-by: Tigran Mkrtchyan