dcache-view (admin): add quota-info-view (read-only
Motivation:

Allow quotas to be listed in dCacheView as part of the
admin interface.

Modfication:

Add the view, consisting of a table with sorters and
filters which merges user and group quota info.

Add icon on the admin toolbar.

Result:

Read-only listing of quotas.

Target: master
Patch: https://rb.dcache.org/r/13137/
Depends-on: #13136
Acked-by: Marina