dcache-view: change User Profile to use Google pie chart for quota info
Motivation:

Better visual representation for user.

Modification:

Quotas have two headings, the main user quota (in dark green)
and the user's group quotas (in dark blue).

Currently, only two quota types are supported (CUSTODIAL, REPLICA),
so the grid size is a constant 2.

If the quota is not defined for one of them, "Quota not defined"
will appear in that slot.

If there is no quota defined for an id, no entry will appear at all.

The pie charts are magnifiable via tap/click, as usual.  When under
quota, the pie chart shows the used and remaining in percentages,
and hover-over the actual number of bytes.  When over quota,
the pie chart switches to red, with the slice showing the part
of the total (quota limit + excess) occupied by the limit (red) and
excess (dark red).  Again, hover-over gives the bytes.

Result:

More GUI-idiomatic, hopefully.  (Note, the full listing for the
admin of all quotas is still in table form.)

Target: master
Patch: https://rb.dcache.org/r/13159/
Requires-notes: yes
Requires-book: no
Acked-by: Tigran
Acked-by: Lea