dcache-view: add listing virtual directories(labels)
Modifications

There is a lot of redundant code, which should be optimized.

a new  file-metadata-task-labels.js is added. the reson is that the ls is done on "labels" not "namespace"
(${endpoint}labels${path})

There is a change in geting the fulPath for delete and setLabel finctions, because the virtual dir the file name includes the parent as well.