shareable-file-generation: POST to file location, not /
Creating a Macaroon by doing POST to / has been proven very fragile in
our setups, both production and test environments. If something else is
handling /, such as a static page, the POST fails with 500 Server Error
without much helpful info in the logs.

Changing this to POST:ing to the file location works around this
problem, this is also closer to what is done to other Macaroon-using
systems such as Rucio.

As POST:ing to the file location makes dCache add a path caveat, avoid
adding another redundant one.

Signed-off-by: Niklas Edmundsson <nikke@hpc2n.umu.se>