file share: Display direct link and rclone config snippet
Display direct link (including QR code) to shared object.

If the shared object is a directory, also includes an rclone config
snippet for easy access using the rclone utility, as documented on
https://rclone.org/webdav/#dcache

QR codes are now generated in a more proper size instead of being scaled
down, this makes it easier for QR reader apps to parse them.

*NOTE:* The related layout fixes are very crude, and likely needs to be
reworked by someone more proficient in the framework and HTML/CSS
interaction.

In order to give the user more context, the share dialogs print the
complete object name together with information on whether it's a file or
directory.

Also, help tooltips are added to the success dialog.

In addition there are also minor fixes related to spelling/layout etc.

Signed-off-by: Niklas Edmundsson <nikke@hpc2n.umu.se>