file-list: Also display unix mode/permissions
Users sometimes wants to know the mode (permission) flags of objects
without having to show the file details of each individual item.

This patch adds the familiar unix ls-style representation to the file
listing.

In order to closer resemble the familiar/expected style, the common
function convertToNameValueObject() is also augmented with an optional
parameter that allows displaying the file type instead of ? since the
mode bits available doesn't include this information.

Fixes: https://github.com/dCache/dcache-view/issues/202

Signed-off-by: Niklas Edmundsson <nikke@hpc2n.umu.se>