Sort file view by file name
Currently the file view is unsorted, which causes great frustration
among users even when dealing with a small amount of items.

Since the Polymer iron-list doesn't seem to have a built-in
functionality the general method of tackling this seems to be to sort
the input data, which matches badly with the chunked approach of
obtaining the file listing from the webdav endpoint.  In order to
quickly get a solution in place that placates users, this is a
bare-minimum approach that increases the file listing chunk size from
100 to 10000 items, and sorts each chunk before adding it to the
iron-list file listing.

The result of this being that file listings are sorted if they contain
up to 10000 items, for listings larger than the file listings will look
a bit weird with groupings of sorted chunks. Quick testing suggests that
it's the listing operation that takes time and not the sorting, so the
chunk size could likely be increased even more.

Fixes: https://github.com/dCache/dcache-view/issues/285

Signed-off-by: Niklas Edmundsson <nikke@hpc2n.umu.se>