dcache-view:  eliminate role assertion
Motivation:

With #14083, the frontend services no longer use the
roles in the LoginAttributes supplied through the
RolesPlugin, but rely on the presence of
the RolePrincipal in the user's Subject.

This allows us to avoid the extra login to get
admin privileges.

Modification:

Eliminate all code having to do with role request
and assertion.

Display whatever roles automatically come along
with the user's authentication from the cert
(or OIDC) after the uid and gids in the user
profile.

Result:

No longer is there any need to do a user-password
login to obtain admin privileges.  These are
automatically accorded to the user according
to the authorization mappings.   This
also applies to the Swagger pages.

Target: master
Patch: https://rb.dcache.org/r/14084/
Requires-notes: yes
Acked-by: Tigran